sw(M5): XLEN-clean lib headers; rv64 default-app link and boot fixes
csr.h/trap.h/limits.h carried ilp32 assumptions that go wrong at lp64:
csr_read and friends returned uint32_t (truncating 64-bit CSR values),
MCAUSE_INTERRUPT_BIT was pinned at bit 31 (the interrupt bit is
XLEN-1), set_trap_handler cast the handler pointer through uint32_t
(truncating the mtvec target), and LONG_MIN/LONG_MAX/ULONG_MAX were
32-bit at an ABI where long is 64. The accessors now return
XLEN-natural unsigned long (identical width and codegen at ilp32),
mcause/SD constants key off __riscv_xlen, mstatus gains the FS field
constants (D15), the interrupt save/restore pair and trap-handler
accessors are pointer/XLEN clean, and limits.h selects long limits by
ABI. The Zicntr high-half readers (rdcycleh/rdtimeh/rdinstreth) are
now RV32-only -- those CSRs do not exist at RV64 and trap since M5.1
-- and rdcycle64/rdtime64/rdinstret64 read the single full-width CSR
at 64 instead of the high/low wrap loop.

Two pre-existing rv64 gaps surfaced by building the default app
config at 64 (never linked before; the rv64 lanes all use bespoke
flows): common.mk now applies -mcmodel=medany at lp64 (medlow's
absolute lui cannot form 0x8xxx_xxxx addresses; same treatment as the
riscv_tests/arch_test/reference builds), and the shared crt0 loads
the .ddr_bss bounds from a local literal pool of absolute dwords at
64 -- la's auipc cannot span the 2 GiB from the low-BRAM .init to the
cached region (the established lp64-reach class). The rv32 branch of
both is byte-identical to before.

Verified: all 62 apps compile at rv32 (the four non-compiling ones
require TEST_SRC-style variables by design and do not include these
headers); hello_world boots and passes at rv32 AND at rv64 (the first
default-config rv64 app boot); amo_irq_torture (DDR-boot, exercises
the widened trap.h interrupt path and MCAUSE_INTERRUPT_BIT at
runtime) passes at its documented sim scale; umode_test green.

+++ sw/lib/include/trap.h
@@ -110,9 +110,9 @@ static inline __attribute__((always_inline)) void enable_interrupts(void)
  *
  * Returns the previous mstatus value so it can be restored later.
  */
-static inline __attribute__((always_inline)) uint32_t disable_interrupts(void)
+static inline __attribute__((always_inline)) unsigned long disable_interrupts(void)
 {
-    uint32_t prev = csr_read(mstatus);
+    unsigned long prev = csr_read(mstatus);
     csr_clear(mstatus, MSTATUS_MIE);
     return prev;
 }
@@ -120,7 +120,7 @@ static inline __attribute__((always_inline)) uint32_t disable_interrupts(void)
 /**
  * Restore interrupt state from a previous disable_interrupts() call
  */
-static inline __attribute__((always_inline)) void restore_interrupts(uint32_t mstatus_val)
+static inline __attribute__((always_inline)) void restore_interrupts(unsigned long mstatus_val)
 {
     if (mstatus_val & MSTATUS_MIE) {
         csr_set(mstatus, MSTATUS_MIE);
@@ -192,13 +192,14 @@ static inline __attribute__((always_inline)) void disable_external_interrupt(voi
  */
 static inline void set_trap_handler(void (*handler)(void))
 {
-    csr_write(mtvec, (uint32_t) handler);
+    /* uintptr_t: a uint32_t cast would truncate the handler address at RV64 */
+    csr_write(mtvec, (uintptr_t) handler);
 }
 
 /**
  * Get the current trap handler address
  */
-static inline uint32_t get_trap_handler(void)
+static inline uintptr_t get_trap_handler(void)
 {
     return csr_read(mtvec);
 }
